# chapter13/whatstore/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Whatstore API",
        "version": "0.1"
    },
    "paths": {
...
        "/inventoryapi/employeeapi/{key}": {
            "get": {
                "operationId": "d9eb586c9794fe690e0e40c1b53a4cc1",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee ID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get an employee"
                    }
                }
            },
            "put": {
                "operationId": "df8e9bffcd78b0afbe88b23eae2905a5",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee ID",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update an employee"
                    }
                }
            },
            "post": {
                "operationId": "db3d93429b54a2775993b3b14fdb4b8c",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee ID",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "ID": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert an employee"
                    }
                }
            },
            "delete": {
                "operationId": "638606c6567bde37d3e4b7df93439ca4",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee ID",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove an employee"
                    }
                }
            }
        },
...
        "/inventoryapi/employeeroleapi/{key}": {
            "post": {
                "operationId": "1d995e74784942ddd0ecfcda2a9a31f4",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee role code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code_role",
                                    "ID_employee"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "integer"
                                    },
                                    "code_role": {
                                        "type": "integer"
                                    },
                                    "ID_employee": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "insert a employee role"
                    }
                }
            },
            "delete": {
                "operationId": "2ca8eb4b44d330c44886830aa18d4cd3",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "employee role code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "remove a employee role"
                    }
                }
            }
        },
        "/inventoryapi/inventoryapi/{key}": {
            "get": {
                "operationId": "ccf828b552e0ddd9347fd4091323c71c",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get an inventory"
                    }
                }
            },
            "put": {
                "operationId": "43a6b5362f2a680ece74b5154fda8525",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "product code",
                        "required": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "code_product": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "operation": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "update an inventory"
                    }
                }
            }
        },
...
    }
}